Header and footer tidy
Added the emacs mode magic to the first line to ensure that the files are
always treated as C files in the elvin style.  Canonicalized the copyright
statement and updated the year to include 2009.  Added a footer to C files
so that there's always a clear last line of the file.