Streamline message allocation.
Messages are mostly immutable; once they're created with message_alloc,
their string contents never change.  We can take advantage of this property
and allocate the message along with room for each of its strings in a single
block of memory, reducing heap fragmentation and improving locality of
reference.  Plus it's nifty.