Add an Edit menu.
Create a new Edit menu to hold Copy and Copy Link menu items which should
make it easier for the user to extract useful information from this
application.

First stab at a copying the Motif way.

Copy the selected message to the clipboard.  This uses the Motif-1.2
mechanism and doesn't work so well with xterm, though it does if you
bounce things through xclipboard.  It still sucks.