Resolved a naming conflict.
In alloc_font_code_set, we refer to two separate arrays of atoms as 'atoms'.
Originally the name only referred to the array of atoms passed to the
XGetAtomNames call to look up the registry and encoding of a font.  We now
have a global atoms array, and we get the Atoms we use to populate this
array from that one.  Use a different name for the old array.
 between the global atoms array and that used
by the code which