Mostly re-instate David's original keybindings.
In David's original patch, <Ctrl>c copied the message text to the
primary selection and <Shift>c copied the message id to the same
location.  The current consensus on the use of the primary selection
is that it should only hold text selected via the mouse pointer and
that the clipboard should be generally preferred, so we copy to that
destination instead.