Clean up message_debug.
It's only used from within message.c, so make it a static function.
Activate it via a macro, like we do with DPRINTF and use those
functions consistently.  Arrange for the entire message to be dumped
in a single call to DPRINTF and, finally, avoid printing double-quotes
around NULL values to reduce clutter.