Be more careful when replacing glyphs
A glyph can easily be visible even if its final successor is not.
Walk the list of glyph holders to find (one of) the glyph(s) being
replaced and replace its non-visible successor with the new glyph if
it has one.  This ensures that the successor list is never more than
three elements long (up to two visible glyphs and an as-yet-unseen
successor).